App/reddit_to_kafka/requirements.txt
kafka_python==2.0.2
praw==7.8.1
python-dotenv==1.0.1
